Refactor App.jsx to Home component
